**fixed mail.ttd \n **fixed compile error in source/VISUAL/SpectraIdentificationViewWidget.C
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@7585 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/VISUAL/SpectraIdentificationViewWidget.C
@@ -217,7 +217,7 @@ namespace OpenMS
           // index
           item = new QTableWidgetItem();
           item->setTextAlignment(Qt::AlignCenter);
-          item->setData(Qt::DisplayRole, i);
+          item->setData(Qt::DisplayRole, UInt(i));
           item->setBackgroundColor(c);
           table_widget_->setItem(table_widget_->rowCount()-1 , 1, item);
